+++ const_format/src/lib.rs
@@ -4,8 +4,7 @@
 //!
 //! # Rust versions
 //!
-//! There are some features that require a variety of stable Rust versions and
-//! others that require Rust nightly,
+//! There are some features that require a variety of Rust versions,
 //! the sections below describe the features that are available for each version.
 //!
 //! ### Rust 1.57.0
@@ -53,12 +52,11 @@
 //!
 //! -  [`str_split`]: splits a string constant
 //!
-//! ### Rust nightly
+//! ### Rust 1.83.0
 //!
 //! By enabling the "fmt" feature, you can use a [`std::fmt`]-like API.
 //!
-//! This requires the nightly compiler, because it uses mutable references in const fn,
-//! which have not been stabilized as of writing these docs.
+//! This requires Rust 1.83.0, because it uses mutable references in const fn.
 //!
 //! All the other features of this crate are implemented on top of the [`const_format::fmt`] API:
 //!
@@ -115,11 +113,11 @@
 //! This example demonstrates how you can use the [`ConstDebug`] derive macro,
 //! and then format the type into a `&'static str` constant.
 //!
-//! This example requires Rust nightly, and the `"derive"` feature.
+//! This example requires the `"derive"` feature
+//! and Rust 1.83.0, because it uses `&mut` in a const context.
 //!
 #![cfg_attr(feature = "derive", doc = "```rust")]
 #![cfg_attr(not(feature = "derive"), doc = "```ignore")]
-//! #![feature(const_mut_refs)]
 //!
 //! use const_format::{ConstDebug, formatc};
 //!
@@ -245,15 +243,15 @@
 //!
 //! # Cargo features
 //!
-//! - `"fmt"`: Enables the [`std::fmt`]-like API,
-//! requires Rust nightly because it uses mutable references in const fn.<br>
+//! - `"fmt"`: Enables the [`std::fmt`]-like API and `"rust_1_83"` feature,
+//! requires Rust 1.83.0 because it uses mutable references in const fn.<br>
 //! This feature includes the [`formatc`]/[`writec`] formatting macros.
 //!
-//! - `"derive"`: requires Rust nightly, implies the `"fmt"` feature,
+//! - `"derive"`: requires Rust 1.83.0, implies the `"fmt"` feature,
 //! provides the [`ConstDebug`] derive macro to format user-defined types at compile-time.<br>
 //! This implicitly uses the `syn` crate, so clean compiles take a bit longer than without the feature.
 //!
-//! - `"assertc"`: requires Rust nightly, implies the `"fmt"` feature,
+//! - `"assertc"`: requires Rust 1.83.0, implies the `"fmt"` feature,
 //! enables the [`assertc`], [`assertc_eq`], and [`assertc_ne`] assertion macros.<br>
 //! This feature was previously named `"assert"`,
 //! but it was renamed to avoid confusion with the `"assertcp"` feature.
@@ -264,6 +262,9 @@
 //! - `"rust_1_64"`: Enables the [`str_split`] macro.
 //! Allows the `as_bytes_alt` methods and `slice_up_to_len_alt` methods to run
 //! in constant time, rather than linear time (proportional to the truncated part of the slice).
+//!
+//! - `"rust_1_83"`: Enables the `"rust_1_64"` feature
+//! and makes macros that evaluate to a value compatible with [inline const patterns].
 //!
 //! # No-std support
 //!
@@ -336,8 +337,9 @@
 //!
 //! [`str::replace`]: https://doc.rust-lang.org/std/primitive.str.html#method.replace
 //!
+//! [inline const patterns]: https://doc.rust-lang.org/1.83.0/unstable-book/language-features/inline-const-pat.html
+//!
 #![no_std]
-#![cfg_attr(feature = "fmt", feature(const_mut_refs))]
 #![cfg_attr(feature = "__docsrs", feature(doc_cfg))]
 #![deny(rust_2018_idioms)]
 // This lint is silly
@@ -488,6 +490,12 @@ pub mod pmr {
     #[repr(transparent)]
     pub struct __AssertStr {
         pub x: &'static str,
+    }
+
+    #[doc(hidden)]
+    #[repr(transparent)]
+    pub struct __AssertType<T> {
+        pub x: T,
     }
 }