+++ const_format/tests/misc_tests/assertcp_tests.rs
@@ -1,4 +1,5 @@
 #![allow(unreachable_code)]
+#![allow(non_local_definitions)]
 
 use cfmt_b::{assertcp, assertcp_eq, assertcp_ne};